--- pyincore/analyses/buildingportfolio/__init__.py
@@ -1,3 +1,0 @@
-from pyincore.analyses.buildingportfolio.BuildingPortfolioRecoveryAnalysis import BuildingPortfolioRecoveryAnalysis
-from pyincore.analyses.buildingportfolio.BuildingDamage import BuildingDamage
-from pyincore.analyses.buildingportfolio.BuildingData import BuildingData
